comments & refine
